added draw to circle

+++ utils.h
@@ -30,7 +30,7 @@ string stringify(const T& var){
 };
 
 string psLine(int,int,int,int);
-string psArc(int,int,int,int,int);
+string psArc(int,int,double,int,int);
 
 
 #endif
